Refactor content_query editing into ViewSmartSearchDialog component
